% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigassertr-package.R
\docType{package}
\name{bigassertr-package}
\alias{bigassertr}
\alias{bigassertr-package}
\title{bigassertr: Assertions}
\description{
Assertion and message functions.
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
