% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{printf}
\alias{printf}
\alias{message2}
\alias{warning2}
\alias{stop2}
\title{Easy messages}
\usage{
printf(...)

message2(...)

warning2(...)

stop2(...)
}
\arguments{
\item{...}{Arguments passed on to \link[base:sprintf]{base::sprintf}.}
}
\description{
Easy messages
}
\examples{
printf("My name is \%s.", "Florian")
}
\seealso{
\link[base:sprintf]{base::sprintf}
}
